+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_de.dart
@@ -666,4 +666,10 @@ class StreamChatLocalizationsDe extends GlobalStreamChatLocalizations {
 
   @override
   String get draftLabel => 'Entwurf';
+
+  @override
+  String locationLabel({bool isLive = false}) {
+    if (isLive) return '📍 Live-Standort';
+    return '📍 Standort';
+  }
 }