+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_en.dart
@@ -669,4 +669,10 @@ class StreamChatLocalizationsEn extends GlobalStreamChatLocalizations {
 
   @override
   String get draftLabel => 'Draft';
+
+  @override
+  String locationLabel({bool isLive = false}) {
+    if (isLive) return '📍 Live Location';
+    return '📍 Location';
+  }
 }