+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_pt.dart
@@ -674,4 +674,10 @@ Não é possível adicionar mais de $limit arquivos de uma vez
 
   @override
   String get draftLabel => 'Rascunho';
+
+  @override
+  String locationLabel({bool isLive = false}) {
+    if (isLive) return '📍 Localização ao Vivo';
+    return '📍 Localização';
+  }
 }