+++ packages/stream_chat_persistence/lib/src/dao/message_dao.dart
@@ -63,7 +63,7 @@ class MessageDao extends DatabaseAccessor<DriftChatDatabase>
     final draft = await switch (fetchDraft) {
       true => _db.draftMessageDao.getDraftMessageByCid(
           msgEntity.channelCid,
-          parentId: msgEntity.parentId,
+          parentId: msgEntity.id,
         ),
       _ => null,
     };