+++ packages/stream_chat_persistence/lib/src/dao/pinned_message_dao.dart
@@ -62,7 +62,7 @@ class PinnedMessageDao extends DatabaseAccessor<DriftChatDatabase>
     final draft = await switch (fetchDraft) {
       true => _db.draftMessageDao.getDraftMessageByCid(
           msgEntity.channelCid,
-          parentId: msgEntity.parentId,
+          parentId: msgEntity.id,
         ),
       _ => null,
     };